+++ src/Sider/RedisClient.IO.core.cs
@@ -12,20 +12,43 @@ namespace Sider
     private Queue<Func<RedisReader, object>> _readsQueue;
 
 
-    private IEnumerable<object> pipelineCore(Action<IRedisClient<T>> pipelinedCalls)
+    private IEnumerable<object> executePipeline(Action<IRedisClient<T>> pipelinedCalls,
+      int retryCount = 0)
     {
-      _readsQueue = _readsQueue ?? new Queue<Func<RedisReader, object>>();
+      object[] results;
 
-      // all writes executed immediately but reads are queued
-      _isPipelining = true;
-      pipelinedCalls(this);
-      _writer.Flush();
-      _isPipelining = false;
+      try {
+        _readsQueue = _readsQueue ?? new Queue<Func<RedisReader, object>>();
+        _readsQueue.Clear();
+
+        // all writes executed immediately but reads are queued
+        _isPipelining = true;
+        pipelinedCalls(this);
+        _writer.Flush();
+        _isPipelining = false;
 
 
-      // reads out all the return values
-      while (_readsQueue.Count > 0)
-        yield return _readsQueue.Dequeue()(_reader);
+        // reads out all the return values
+        results = new object[_readsQueue.Count];
+        var resultIdx = 0;
+
+        while (_readsQueue.Count > 0)
+          results[resultIdx++] = _readsQueue.Dequeue()(_reader);
+      }
+      catch (Exception ex) {
+        if (!handleException(ex))
+          throw;
+
+        // TODO: multiple retries with pipeline maybe too dangerous
+        //   so will only retry once in pipeline mode.
+        if (_settings.ReissuePipelinedCallsOnReconnect &&
+          retryCount < 1)
+          return executePipeline(pipelinedCalls, retryCount + 1);
+
+        throw;
+      }
+
+      return results;
     }
 
 
@@ -38,27 +61,22 @@ namespace Sider
     {
       try { return func(); }
       catch (Exception ex) {
-        if (!handleException(ex))
+        if (_isPipelining || !handleException(ex))
           throw;
 
-        // exception processed by this point
-        if ((_settings.ReissueWriteOnReconnect && ex is WriteException) ||
-          (_settings.ReissueReadOnReconnect && ex is ReadException)) {
+        // acceptable exception processed by this point
+        // retry if allowed to -- usually harmless since if there's a
+        // disconnection it usually means nothing was processed on Redis side.
+        if (_settings.ReissueCommandsOnReconnect &&
+          retryCount < _settings.MaxReconnectRetries)
+          return execute(func, retryCount + 1);
 
-          // retry once, if allowed to -- usually harmless even with writes
-          // since if there's a disconnection it should means nothing was processed.
-          if (retryCount < _settings.MaxReconnectRetries)
-            return execute(func, retryCount + 1);
-
-          // no more retries
-          throw;
-        }
-
-        // signal the client that the action failed.
-        // (even though we've reconnected)
+        // signal the client that the action failed or retried too many times
+        // even if we've reconnected
         throw;
       }
     }
+
 
     private bool handleException(Exception ex)
     {
@@ -73,16 +91,11 @@ namespace Sider
       // assumes Redis disconnected us due to timeouts
       Dispose();
       if (!_settings.ReconnectOnIdle)
-        throw timeoutException(ex);
+        throw new TimeoutException(
+          "Disconnection detected. Probably due to idle timeout.", ex);
 
       Reset();
       return true;
-    }
-
-    private Exception timeoutException(Exception inner)
-    {
-      return new TimeoutException(
-        "Disconnection detected. Probably due to idle timeout.", inner);
     }